ops-manage-back/requirements.txt
concurrent-log==1.0.1
coreapi==2.3.3
coreschema==0.0.4
dataclasses
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
apscheduler==3.9.1
#swagger
drf-yasg==1.20.0
swagger-spec-validator==2.7.3
PyMySQL==0.9.3
gevent==21.1.2
gunicorn==20.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
python-dotenv==0.17.1
python-ldap==3.3.1
PyYAML==5.4.1
requests==2.27.0
user-agents==2.2.0
uuid==1.30
PyJWT==2.1.0
whitenoise==5.2.0
django-simple-captcha==0.5.17
Pillow==9.5
channels==3.0.3
uvicorn==0.14.0
elasticsearch==8.3.3
pandas==1.1.5
kubernetes==25.3.0
celery==5.1.2
cryptography==2.4.2
redis==4.3.3
lxml==4.9.3